Rename clipper lib to polyclipping
The library is called polyclipping and both the Ubuntu and Arch Linux
packages are also called polyclipping.

Renaming the directory allows for easier unbundling of the library.
