+++ ci/stylecheck.sh
@@ -4,13 +4,13 @@
 set -euv -o pipefail
 
 # check if all files have Unix line endings (except 3rd-party libs)
-(git grep -Il $'\r' -- ':/' ':!/libs/clipper/') && exit 1
+(git grep -Il $'\r' -- ':/' ':!/libs/polyclipping/') && exit 1
 
 # check if no file contains trailing spaces (with some exceptions)
 (git grep -Il ' $' -- ':/' ':!/share/librepcb/licenses/' ':!*.dia') && exit 1
 
 # check if no file contains tabulators (with some exceptions)
-(git grep -Il $'\t' -- ':/' ':!/libs/clipper/' ':!/share/librepcb/licenses/' ':!*.ui') && exit 1
+(git grep -Il $'\t' -- ':/' ':!/libs/polyclipping/' ':!/share/librepcb/licenses/' ':!*.ui') && exit 1
 
 # run python style checks
 (flake8 --ignore=E501 dev tests) || exit 1
