+++ libs/librepcb/project/boards/drc/boarddesignrulecheck.h
@@ -25,7 +25,7 @@
  ******************************************************************************/
 #include "boarddesignrulecheckmessage.h"
 
-#include <clipper/clipper.hpp>
+#include <polyclipping/clipper.hpp>
 
 #include <QtCore>
