+++ libs/polyclipping/License.txt
