+++ libs/polyclipping/README
