+++ libs/polyclipping/clipper.cpp
