+++ libs/polyclipping/clipper.hpp
